Added stores, now accessible in route files
There are two types of stores: `session` and `persistent`

The goal of the stores is to keep data between requests, making output
more predictable when random data is generated and kept for later use.

Two new methods have been added to the route sandbox:
* `getSessionStore()`
* `getPersistentStore()`

Flat file persistent storage is unique to the directory of the active
scenarios file. The `dirty` module provides persistant storage.